requirements.txt
# Core dependencies
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.4.0
numpy>=1.26.4,<2.0

# Embedding & Search
sentence-transformers>=3.4.0
transformers>=4.48.0
faiss-cpu==1.8.0
sentencepiece>=0.2.0
protobuf>=5.29.0

# PDF Processing
PyMuPDF==1.24.0

# API Server
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# CLI
click==8.1.7
requests==2.31.0

# Utilities
tqdm==4.66.2
watchdog==4.0.0
